Airbyte/requirements.txt
pyyaml==6.0.3
loguru==0.7.3
requests==2.32.5
